// apps/server/src/modules/common-cartridge/common-cartridge-client/room-client/room-api-client/models/board-lesson-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface BoardLessonResponse
 */
export interface BoardLessonResponse {
    /**
     * 
     * @type {string}
     * @memberof BoardLessonResponse
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof BoardLessonResponse
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof BoardLessonResponse
     */
    'courseName'?: string;
    /**
     * 
     * @type {number}
     * @memberof BoardLessonResponse
     */
    'numberOfPublishedTasks': number;
    /**
     * 
     * @type {number}
     * @memberof BoardLessonResponse
     */
    'numberOfDraftTasks': number;
    /**
     * 
     * @type {number}
     * @memberof BoardLessonResponse
     */
    'numberOfPlannedTasks': number;
    /**
     * 
     * @type {string}
     * @memberof BoardLessonResponse
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof BoardLessonResponse
     */
    'updatedAt': string;
    /**
     * 
     * @type {boolean}
     * @memberof BoardLessonResponse
     */
    'hidden': boolean;
}